move modules to subfolder
